gateway-core/src/main/java/io/conduktor/gateway/config/kafka/FileKafkaPropertiesProvider.java
/*
 * Copyright 2023 Conduktor, Inc
 *
 * Licensed under the Conduktor Community License (the "License"); you may not use
 * this file except in compliance with the License.  You may obtain a copy of the
 * License at
 *
 * https://www.conduktor.io/conduktor-community-license-agreement-v1.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations under the License.
 */

package io.conduktor.gateway.config.kafka;

import lombok.extern.slf4j.Slf4j;

import java.io.FileInputStream;
import java.io.IOException;
import java.io.InputStream;
import java.util.Properties;

@Slf4j
public record FileKafkaPropertiesProvider(String configFile) implements KafkaPropertiesProvider {

    @Override
    public Properties loadKafkaProperties() throws IOException {
        log.info("Loading Kafka Config from file: " + configFile);
        final Properties cfg = new Properties();
        try (InputStream inputStream = new FileInputStream(configFile)) {
            cfg.load(inputStream);
        } catch (IOException e) {
            throw new IOException("Error loading kafka configuration properties from file " + configFile, e);
        }
        return cfg;
    }

}
